Refactor media insertion code to an helper class.
